+++ lib/decorators.d.ts
@@ -1,2 +1,4 @@
 export declare function Required(): MethodDecorator;
 export declare function Requires(methodName: string | symbol): MethodDecorator;
+export declare function AuthorizedAttribute(attribute: string | string[]): MethodDecorator;
+export declare function AuthorizedResource(resource: string): MethodDecorator;