+++ lib/errors.d.ts
@@ -1,2 +1,3 @@
-export declare class RFIAuthError {
+import { RFIError } from '@rumbleship/spyglass';
+export declare class RFIAuthError extends RFIError {
 }