+++ src/authorizer.ts
@@ -18,6 +18,7 @@ export class Authorizer {
   private client?: string;
   private roles?: RolesAt;
   private scopes?: Scopes[] = [];
+  private owner?: string;
 
   constructor(private authorizationHeader: string, private secret: string) {
     if (!this.authorizationHeader) {
@@ -36,11 +37,15 @@ export class Authorizer {
 
   @Required()
   authenticate(): boolean {
-    const { roles, scopes, user, client } = jwt.verify(this.accessToken, this.secret) as Claims;
+    const { roles, scopes, user, client, owner } = jwt.verify(
+      this.accessToken,
+      this.secret
+    ) as Claims;
     this.user = user;
     this.roles = roles;
     this.scopes = scopes;
     this.client = client;
+    this.owner = owner;
     return !!this.roles;
   }
 
@@ -58,6 +63,17 @@ export class Authorizer {
   getClient(): string | undefined {
     return this.client;
   }
+  @Requires('authenticate')
+  getOwner(): string | undefined {
+    // this really should return an unwrappable Oid...
+    return this.owner;
+  }
+
+  @Requires('authenticate')
+  getAuthorizationHeader(): string {
+    return this.authorizationHeader;
+  }
+  // I think this is leftover cruft? Would like to remove.
   getClaims(): Claims {
     return jwt.verify(this.accessToken, this.secret) as Claims;
   }
@@ -78,7 +94,7 @@ export class Authorizer {
     action: Actions,
     authorizable: object,
     matrix: PermissionsMatrix[],
-    attribute?: string,
+    attribute?: string | string[],
     resource?: Resource
   ) {
     let access = false;
@@ -99,11 +115,14 @@ export class Authorizer {
           const actions = resource
             ? (group as any)[resource] || []
             : (group as any)[authorizable.constructor.name] || [];
-          if (
-            permissionedIdentifiers.includes((authorizable as any)[attribute]) &&
-            (actions as any).includes(action)
-          ) {
-            access = true;
+          const attrs = !Array.isArray(attribute) ? [attribute] : attribute;
+          for (const attr of attrs) {
+            if (
+              permissionedIdentifiers.includes((authorizable as any)[attr]) &&
+              (actions as any).includes(action)
+            ) {
+              access = true;
+            }
           }
         }