+++ src/types.ts
@@ -1,3 +1,4 @@
+import { RefreshClaims } from './types';
 export enum Roles {
   ADMIN = 'admin',
   USER = 'user',
@@ -13,7 +14,11 @@ export enum Actions {
   // are Approve and Verify the same action, e.g. a limited update?
   APPROVE = 'approve',
   VERIFY = 'verify',
-  REQUEST = 'request'
+  REQUEST = 'request',
+  // these values are uppercased to be compatible with keys to make it easier to access values without having to change the case --2019-08-06 @preeV42500
+  PURCHASE = 'PURCHASE',
+  SHIP = 'SHIP',
+  RETURN = 'RETURN'
 }
 export enum Resource {
   Division = 'Division',
@@ -24,7 +29,13 @@ export enum Resource {
 export enum Scopes {
   SYSADMIN = 'system:*',
   BANKINGADMIN = 'banking:*',
-  ORDERADMIN = 'orders:*'
+  ORDERADMIN = 'orders:*',
+  DIVISIONADMIN = 'divisions:*'
+}
+
+export enum GrantTypes {
+  REFRESH = 'refresh',
+  ACCESS = 'access'
 }
 
 export type PermissionsGroup = {
@@ -42,14 +53,18 @@ export type RolesAt = {
   [key in Roles]?: string[]; // should be Oid[]
 };
 
-export interface ClaimsInput {
+export interface AccessClaims {
   name?: string;
   user?: string;
   client?: string;
   roles: RolesAt;
   scopes: Scopes[];
 }
-export interface Claims extends ClaimsInput {
+export interface RefreshClaims {
+  owner: string;
+  grant_type: GrantTypes;
+}
+export interface Claims extends AccessClaims, RefreshClaims {
   exp: Date;
   iat: Date;
 }