+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file. Starting wi
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+  * DivisionAdmin scope
+  * `.prettierignore` file
+  * Skeleton for generating and consuming `refreshTokens`
+  * Decorators for injecting the `attribute` and `resource` into a subclass's method definition
+  * keys and values in `Actions` enum relating to `OrderEvent`
+### Removed
+  * `moment` is no longer a peer dependency, just a `devDependency`
+### Changed
+  * `Actions.PURCHASE`, `Actions.SHIP` and `Actions.RETURN` values are now uppercase
+  * `can` accepts either an attribute or a list of attributes
+### Fixed
+### Deprecated
+### Security
+
+
 ## [0.0.1] -- 2019-07-17
 
 ### Added